requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.10.0
asgiref==3.9.1
astor==0.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.3.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
Django==5.2.5
django-api-generator==1.0.18
django-debug-toolbar==5.2.0
django-jazzmin==3.0.1
django-mathfilters==1.0.0
django_celery_results==2.6.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
filelock==3.19.1
fsspec==2025.7.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.8
hiredis==3.2.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
inflection==0.5.1
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
str2bool==1.1
tokenizers==0.21.4
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.5.0
